pyproject.toml
[project]
name = "anime-recommender"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.8",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.2",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "sentence-transformers>=5.4.1",
    "streamlit>=1.56.0",
    "torchvision>=0.26.0",
]
